requirements.txt
django==3.1.5
django-ckeditor==6.0.0
psycopg2-binary==2.8.6
uwsgi==2.0.19.1
